+++ Zenith-LanguageModel/src/test/java/com/ciphertool/zenith/model/markov/MarkovModelTest.java
@@ -19,22 +19,24 @@
 
 package com.ciphertool.zenith.model.markov;
 
-import static org.mockito.Mockito.spy;
-
-import java.lang.reflect.Field;
-import java.util.*;
-
 import com.ciphertool.zenith.math.MathConstants;
 import com.ciphertool.zenith.math.sampling.RouletteSampler;
 import com.ciphertool.zenith.model.ModelConstants;
+import com.ciphertool.zenith.model.entities.TreeNGram;
+import com.ciphertool.zenith.model.etl.importers.LetterNGramMarkovImporter;
 import com.ciphertool.zenith.model.probability.LetterProbability;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.util.ReflectionUtils;
 
-import com.ciphertool.zenith.model.entities.TreeNGram;
-import com.ciphertool.zenith.model.etl.importers.LetterNGramMarkovImporter;
+import java.lang.reflect.Field;
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import static org.mockito.Mockito.spy;
 
 public class MarkovModelTest {
 	private Logger								log		= LoggerFactory.getLogger(getClass());
@@ -53,7 +55,7 @@ public class MarkovModelTest {
 		}
 	}
 
-	private static final Float RANDOM_LETTER_TOTAL_PROBABILITY = RANDOM_LETTER_SAMPLER.reIndex(RANDOM_LETTER_PROBABILITIES);
+	private static final BigDecimal RANDOM_LETTER_TOTAL_PROBABILITY = RANDOM_LETTER_SAMPLER.reIndex(RANDOM_LETTER_PROBABILITIES);
 
 	//@BeforeClass
 	public static void setUp() {
@@ -122,7 +124,7 @@ public class MarkovModelTest {
 			probabilities.add(probability);
 		}
 
-		Float totalProbability = sampler.reIndex(probabilities);
+		BigDecimal totalProbability = sampler.reIndex(probabilities);
 
 		int nextIndex = sampler.getNextIndex(probabilities, totalProbability);
